refactor: color,typography 파일로 분리
