+++ packages/theme/typography.ts
@@ -1,0 +1,38 @@
+import { defineTextStyles } from "@pandacss/dev";
+import { typography } from "wowds-tokens";
+
+export const textStyles = defineTextStyles({
+  h1: {
+    value: typography.h1,
+  },
+  h2: {
+    value: typography.h2,
+  },
+  display1: {
+    value: typography.display1,
+  },
+  display2: {
+    value: typography.display2,
+  },
+  h3: {
+    value: typography.h3,
+  },
+  body1: {
+    value: typography.body1,
+  },
+  body2: {
+    value: typography.body2,
+  },
+  body3: {
+    value: typography.body3,
+  },
+  label1: {
+    value: typography.label1,
+  },
+  label2: {
+    value: typography.label2,
+  },
+  label3: {
+    value: typography.label3,
+  },
+});
